+++ Visualizer/audio/prefs.h
@@ -9,9 +9,10 @@ public:
   PWAVEFORMATEX m_pwfx;
   LPCWSTR m_szFilename;
   std::wstring m_szAudioDeviceDisplayName;
+  int m_nAudioDeviceRequesType = 0; // 0: Undefined, 1: Capture (in), 2: Render (out)
 
   // set hr to S_FALSE to abort but return success
-  CPrefs(int argc, LPCWSTR argv[], HRESULT& hr);
+  CPrefs(int argc, LPCWSTR argv[], HRESULT& hr, int nAudioDeviceRequesType);
   ~CPrefs();
 
 };